Part4: Added exercises complete exercises

+++ part4/blog_list/controllers/blogs.js
@@ -3,7 +3,10 @@ const Blog = require("../models/blog");
 
 blogsRouter.get("/", async (_, response, next) => {
   try {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate("user", {
+      username: 1,
+      name: 1,
+    });
     response.json(blogs);
   } catch (error) {
     next(error);
@@ -25,20 +28,49 @@ blogsRouter.get("/:id", async (request, response, next) => {
 
 blogsRouter.post("/", async (request, response, next) => {
   try {
-    console.log(request.body);
-    const blog = new Blog(request.body);
-    console.log("Posting a new blog", blog);
+    const body = request.body;
+    const user = request.user;
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+      user: user.id,
+    });
 
     const result = await blog.save();
+    user.blogs = user.blogs.concat(result._id);
+    await user.save();
+
     response.status(201).json(result);
   } catch (error) {
     next(error);
   }
 });
 
-blogsRouter.delete("/:id", async (request, response, _) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", async (request, response, next) => {
+  try {
+    const user = request.user;
+
+    console.log("user", user);
+
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    console.log("blog", blog);
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response.status(401).json({ error: "not authorized" });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } catch (error) {
+    next(error);
+  }
 });
 
 module.exports = blogsRouter;
